configure, Darwin: Require explicit selection of COBOL.
By defult, Darwin does not have sufficient tools to build COBOL
so we do not want to include it in --enable-languages=all since
this will break regular testing of all supported languages.

However, we do want to be able to build it on demand (where the
build system has sufficiently new tools) and so do not want to
disable it permanently.

ChangeLog:

	* configure: Regenerate.
	* configure.ac: Do not build COBOL on Darwin by default,
	even for --enable-languages=all.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>