cobol: Fix --enable-link-serialization build
--enable-link-serialization relies on each FE participating properly,
setting <lang>.serial, depending on $(<lang>.prev) and printing progress.
The configure option is mainly for LTO bootstraps when we don't want to link
all the FEs at once because that can consume too much memory.

The comment changes are unrelated, just something I've spotted while
working on this.  .exe is a Windows suffix, so either we shouldn't
talk about suffixes in the comments or use there $(exeext) as well
to make it clear that it is dependent on the host/build.

2025-03-11  Jakub Jelinek  <jakub@redhat.com>

	* Make-lang.in: Remove .exe extension from comments.
	(cobol.serial): Set to cobol1$(exeext).
	(cobol1$(exeext)): Depend on $(cobol.prev).  Add
	LINK_PROGRESS calls before/after the link command.