cobol: libgcobol/Makefile.am cleanups
Looking at libgcobol.la, I see a lot of cruft, stuff that just shouldn't
be there because automake generates it otherwise right, but also stuff
using undefined variables etc.
libgcobol.{a,so*} seems to build and install the same as before.

Note, I stull see DT_RUNPATH in the installed libgcobol.so.1 before/after
this patch and I'd prefer not to see it, not seeing it in other libraries
like libstdc++.so.6 etc.  Dunno if that is because of the dependency on
libstdc++ (but e.g. libstdc++ has dependency on libgcc_s and doesn't do
that).

2025-03-11  Jakub Jelinek  <jakub@redhat.com>

	* Makefile.am: Remove tons of VAR = @VAR@ style lines.
	(libgcobol.la): Remove.
	(libgcobol_la_LFLAGS): Remove.
	(all): Remove.
	(.PHONY): Remove.
	(install): Remove.
	(%.lo: %.c): Remove.
	(doc): Remove.
	(install-html install-pdf install-info): Remove.
	* Makefile.in: Regenerate.