arm: testsuite: remove gcc.target/arm/lp1243022.c [PR117931]
This test has been failing since gcc-6.  The test was always very
fragile anyway since it relied on an auto-inc being created and then
split by the subreg2 (later the subreg3) pass.  But the code to get
precisely these conditions was very long-winded and unlikely to be
immune to other changes in the compiler (as proved to be the case).

There's no obvious way to recreate the exact conditions we were
testing for, so just remove the test.

gcc/testsuite:

	PR target/117931
	* gcc.target/arm/lp1243022.c: Delete non-functional test.